feat(agents): unify agent management IA into /agents 3-tab shell (Phase 1)
story d63d3f73 — 통계/관리/채용 탭으로 /agents 를 재구성하고 settings Members에서
에이전트 서브탭을 걷어낸다. 상세는 /agents/[id]로 이동(구 경로는 redirect로 딥링크
보존), /agents/recruiter는 채용 탭으로 redirect. 접근권한 매트릭스(Phase 2)는 BE
bulk endpoint 선행 전까지 미렌더.
