+++ apps/web/messages/en.json
@@ -2377,6 +2377,13 @@
     "recruiterCta": "Recruiter",
     "tabDeployments": "Deployments",
     "tabPerformance": "Performance",
+    "statsTab": "Stats",
+    "manageTab": "Manage",
+    "recruitTab": "Recruit",
+    "scopeThisProject": "This project",
+    "backToList": "Back to list",
+    "manageAddAgent": "Add agent",
+    "manageProjectsGranted": "{count} projects granted",
     "autoBadge": "Auto",
     "autoTooltip": "Created by {name}",
     "createdBy": "Created by {name}",
