+++ apps/web/messages/ko.json
@@ -2377,6 +2377,13 @@
     "recruiterCta": "채용관",
     "tabDeployments": "배포",
     "tabPerformance": "퍼포먼스",
+    "statsTab": "통계",
+    "manageTab": "관리",
+    "recruitTab": "채용",
+    "scopeThisProject": "이 프로젝트",
+    "backToList": "목록으로",
+    "manageAddAgent": "에이전트 추가",
+    "manageProjectsGranted": "{count}개 프로젝트 허용",
     "autoBadge": "Auto",
     "autoTooltip": "{name} 에이전트가 생성함",
     "createdBy": "생성자: {name}",
